Renamed Section to EntitySection
